More granularity in the loss modules. Also, more isort.
