+++ src/diffusion_for_multi_scale_molecular_dynamics/loss/lattice_loss_calculator.py
@@ -1,0 +1,16 @@
+import torch
+
+
+class LatticeLossCalculator(torch.nn.Module):
+    """Class to calculate the loss for the lattice vectors.
+
+    Placeholder for now.
+    """
+
+    def __init__(self):
+        """Placeholder for now."""
+        super().__init__()
+
+    def calculate_unreduced_loss(self, *args):
+        """Placeholder for now."""
+        return 0
