+++ src/diffusion_for_multi_scale_molecular_dynamics/models/instantiate_diffusion_model.py
@@ -3,10 +3,10 @@
 import logging
 from typing import Any, AnyStr, Dict
 
+from diffusion_for_multi_scale_molecular_dynamics.loss.loss_parameters import \
+    create_loss_parameters
 from diffusion_for_multi_scale_molecular_dynamics.models.axl_diffusion_lightning_model import (
     AXLDiffusionLightningModel, AXLDiffusionParameters)
-from diffusion_for_multi_scale_molecular_dynamics.models.loss import \
-    create_loss_parameters
 from diffusion_for_multi_scale_molecular_dynamics.models.optimizer import \
     create_optimizer_parameters
 from diffusion_for_multi_scale_molecular_dynamics.models.scheduler import \
