+++ tests/data/diffusion/test_data_preprocess.py
@@ -4,15 +4,10 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from diffusion_for_multi_scale_molecular_dynamics.data.diffusion.data_preprocess import (
-    LammpsProcessorForDiffusion,
-)
+from diffusion_for_multi_scale_molecular_dynamics.data.diffusion.data_preprocess import \
+    LammpsProcessorForDiffusion
 from diffusion_for_multi_scale_molecular_dynamics.namespace import (
-    ATOM_TYPES,
-    CARTESIAN_FORCES,
-    CARTESIAN_POSITIONS,
-    RELATIVE_COORDINATES,
-)
+    ATOM_TYPES, CARTESIAN_FORCES, CARTESIAN_POSITIONS, RELATIVE_COORDINATES)
 from tests.conftest import TestDiffusionDataBase
 from tests.fake_data_utils import generate_parquet_dataframe
