+++ tests/generators/test_langevin_generator.py
@@ -1,21 +1,16 @@
 import pytest
 import torch
 
-from diffusion_for_multi_scale_molecular_dynamics.generators.langevin_generator import (
-    LangevinGenerator,
-)
-from diffusion_for_multi_scale_molecular_dynamics.generators.predictor_corrector_position_generator import (
-    PredictorCorrectorSamplingParameters,
-)
-from diffusion_for_multi_scale_molecular_dynamics.noise_schedulers.noise_parameters import (
-    NoiseParameters,
-)
-from diffusion_for_multi_scale_molecular_dynamics.utils.basis_transformations import (
-    map_relative_coordinates_to_unit_cell,
-)
-from src.diffusion_for_multi_scale_molecular_dynamics.noise_schedulers.variance_sampler import (
-    NoiseScheduler,
-)
+from diffusion_for_multi_scale_molecular_dynamics.generators.langevin_generator import \
+    LangevinGenerator
+from diffusion_for_multi_scale_molecular_dynamics.generators.predictor_corrector_position_generator import \
+    PredictorCorrectorSamplingParameters
+from diffusion_for_multi_scale_molecular_dynamics.noise_schedulers.noise_parameters import \
+    NoiseParameters
+from diffusion_for_multi_scale_molecular_dynamics.utils.basis_transformations import \
+    map_relative_coordinates_to_unit_cell
+from src.diffusion_for_multi_scale_molecular_dynamics.noise_schedulers.variance_sampler import \
+    NoiseScheduler
 from tests.generators.conftest import BaseTestGenerator
