+++ tests/noise_schedulers/test_variance_sampler.py
@@ -2,12 +2,10 @@ import einops
 import pytest
 import torch
 
-from src.diffusion_for_multi_scale_molecular_dynamics.noise_schedulers.noise_parameters import (
-    NoiseParameters,
-)
-from src.diffusion_for_multi_scale_molecular_dynamics.noise_schedulers.variance_sampler import (
-    NoiseScheduler,
-)
+from src.diffusion_for_multi_scale_molecular_dynamics.noise_schedulers.noise_parameters import \
+    NoiseParameters
+from src.diffusion_for_multi_scale_molecular_dynamics.noise_schedulers.variance_sampler import \
+    NoiseScheduler
 
 
 @pytest.mark.parametrize("total_time_steps", [3, 10, 17])
