+++ tests/sampling/test_diffusion_sampling.py
@@ -3,20 +3,13 @@ import pytest
 import torch
 
 from diffusion_for_multi_scale_molecular_dynamics.namespace import (
-    CARTESIAN_POSITIONS,
-    RELATIVE_COORDINATES,
-    UNIT_CELL,
-)
-from diffusion_for_multi_scale_molecular_dynamics.utils.basis_transformations import (
-    get_positions_from_coordinates,
-)
+    CARTESIAN_POSITIONS, RELATIVE_COORDINATES, UNIT_CELL)
+from diffusion_for_multi_scale_molecular_dynamics.utils.basis_transformations import \
+    get_positions_from_coordinates
 from src.diffusion_for_multi_scale_molecular_dynamics.generators.position_generator import (
-    PositionGenerator,
-    SamplingParameters,
-)
-from src.diffusion_for_multi_scale_molecular_dynamics.sampling.diffusion_sampling import (
-    create_batch_of_samples,
-)
+    PositionGenerator, SamplingParameters)
+from src.diffusion_for_multi_scale_molecular_dynamics.sampling.diffusion_sampling import \
+    create_batch_of_samples
 
 
 class DummyGenerator(PositionGenerator):
@@ -32,7 +25,7 @@ class DummyGenerator(PositionGenerator):
     ) -> torch.Tensor:
         self._counter += number_of_samples
         return self._relative_coordinates[
-            self._counter - number_of_samples : self._counter
+            self._counter - number_of_samples: self._counter
         ]
