+++ tests/test_sample_diffusion.py
@@ -5,24 +5,20 @@ import torch
 import yaml
 
 from diffusion_for_multi_scale_molecular_dynamics import sample_diffusion
-from diffusion_for_multi_scale_molecular_dynamics.generators.predictor_corrector_position_generator import (
-    PredictorCorrectorSamplingParameters,
-)
+from diffusion_for_multi_scale_molecular_dynamics.generators.predictor_corrector_position_generator import \
+    PredictorCorrectorSamplingParameters
+from diffusion_for_multi_scale_molecular_dynamics.loss.loss_parameters import \
+    MSELossParameters
 from diffusion_for_multi_scale_molecular_dynamics.models.axl_diffusion_lightning_model import (
-    AXLDiffusionLightningModel,
-    AXLDiffusionParameters,
-)
-from diffusion_for_multi_scale_molecular_dynamics.models.loss import MSELossParameters
-from diffusion_for_multi_scale_molecular_dynamics.models.optimizer import (
-    OptimizerParameters,
-)
-from diffusion_for_multi_scale_molecular_dynamics.models.score_networks.mlp_score_network import (
-    MLPScoreNetworkParameters,
-)
-from diffusion_for_multi_scale_molecular_dynamics.namespace import RELATIVE_COORDINATES
-from diffusion_for_multi_scale_molecular_dynamics.noise_schedulers.noise_parameters import (
-    NoiseParameters,
-)
+    AXLDiffusionLightningModel, AXLDiffusionParameters)
+from diffusion_for_multi_scale_molecular_dynamics.models.optimizer import \
+    OptimizerParameters
+from diffusion_for_multi_scale_molecular_dynamics.models.score_networks.mlp_score_network import \
+    MLPScoreNetworkParameters
+from diffusion_for_multi_scale_molecular_dynamics.namespace import \
+    RELATIVE_COORDINATES
+from diffusion_for_multi_scale_molecular_dynamics.noise_schedulers.noise_parameters import \
+    NoiseParameters
 
 
 @pytest.fixture()
